+++ tests/CMakeLists.txt
@@ -18,16 +18,19 @@ if(NOT STDLIB_MAP_RECURSIVE)
     target_compile_definitions(testPowerWAF PRIVATE HAS_NONRECURSIVE_UNORDERED_MAP)
 endif()
 
+  #set(GCOV_LINK_FLAGS "-lgcov --coverage")
+  #set(CMAKE_EXE_LINKER_FLAGS "${CMAKE_EXE_LINKER_FLAGS} ${GCOV_LINK_FLAGS}")
+  #set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} -O0 -g --coverage -fprofile-arcs -ftest-coverage")
+
+set(COVERAGE_LINK_FLAGS "")
 if(NOT MSVC)
     if(LIBDDWAF_TEST_COVERAGE)
-        set_target_properties(testPowerWAF PROPERTIES COMPILE_FLAGS "-ggdb --coverage")
-        set(COVERAGE_LINK_FLAGS "--coverage")
+        set_target_properties(testPowerWAF PROPERTIES COMPILE_FLAGS "-ggdb --coverage -fprofile-arcs -ftest-coverage")
+        set(COVERAGE_LINK_FLAGS "-lgcov --coverage")
     else()
         set_target_properties(testPowerWAF PROPERTIES COMPILE_FLAGS "-ggdb")
     endif()
 else()
-  set(COVERAGE_LINK_FLAGS "")
-
   set_target_properties(testPowerWAF PROPERTIES RUNTIME_OUTPUT_DIRECTORY_DEBUG ".")
   set_target_properties(testPowerWAF PROPERTIES RUNTIME_OUTPUT_DIRECTORY_RELEASE ".")
   set_target_properties(testPowerWAF PROPERTIES RUNTIME_OUTPUT_DIRECTORY_MINSIZEREL ".")
