+++ tests/interface_test.cpp
@@ -105,3 +105,50 @@ TEST(TestInterface, HandleLifetimeMultipleContexts)
 
     ddwaf_object_free(&parameter);
 }
+
+TEST(TestInterface, UpdateRules)
+{
+    auto rule = readFile("interface.yaml");
+    ASSERT_TRUE(rule.type != DDWAF_OBJ_INVALID);
+
+    ddwaf_config config{{0, 0, 0}, {nullptr, nullptr}, nullptr};
+
+    ddwaf_handle handle = ddwaf_init(&rule, &config, nullptr);
+    ASSERT_NE(handle, nullptr);
+
+    ddwaf_context context1 = ddwaf_context_init(handle);
+    ASSERT_NE(context1, nullptr);
+
+    ddwaf_handle new_handle = ddwaf_update(handle, &rule, nullptr);
+    ASSERT_NE(new_handle, nullptr);
+
+    ddwaf_object_free(&rule);
+
+    ddwaf_context context2 = ddwaf_context_init(new_handle);
+    ASSERT_NE(context2, nullptr);
+
+    // Destroying the handle should not invalidate it
+    ddwaf_destroy(handle);
+    ddwaf_destroy(new_handle);
+
+    ddwaf_object parameter = DDWAF_OBJECT_MAP;
+    ddwaf_object tmp;
+    ddwaf_object param_key = DDWAF_OBJECT_ARRAY;
+    ddwaf_object param_val = DDWAF_OBJECT_ARRAY;
+
+    ddwaf_object_array_add(&param_key, ddwaf_object_unsigned(&tmp, 4242));
+    ddwaf_object_array_add(&param_key, ddwaf_object_string(&tmp, "randomString"));
+
+    ddwaf_object_array_add(&param_val, ddwaf_object_string(&tmp, "rule1"));
+
+    ddwaf_object_map_add(&parameter, "value1", &param_key);
+    ddwaf_object_map_add(&parameter, "value2", &param_val);
+
+    EXPECT_EQ(ddwaf_run(context1, &parameter, nullptr, LONG_TIME), DDWAF_MATCH);
+    ddwaf_context_destroy(context1);
+
+    EXPECT_EQ(ddwaf_run(context2, &parameter, nullptr, LONG_TIME), DDWAF_MATCH);
+    ddwaf_context_destroy(context2);
+
+    ddwaf_object_free(&parameter);
+}
